Move tests around to the integration file
