+++ async/amqp/integration_test.go
@@ -115,6 +115,42 @@ func TestConsumeFailures(t *testing.T) {
 	}
 }
 
+func TestConsumeAndCancel(t *testing.T) {
+	f := &Factory{
+		url:      "amqp://guest:guest@localhost/",
+		queue:    "queue",
+		exchange: *validExch,
+		bindings: []string{},
+	}
+	c, err := f.Create()
+	require.NoError(t, err)
+	ctx := context.Background()
+	ctx, cancel := context.WithCancel(ctx)
+
+	msgChan, errChan, err := c.Consume(ctx)
+	cancel()
+	assert.Empty(t, msgChan)
+	assert.Empty(t, errChan)
+	assert.NoError(t, err)
+}
+
+func TestConsumeAndClose(t *testing.T) {
+	f := &Factory{
+		url:      "amqp://guest:guest@localhost/",
+		queue:    "queue",
+		exchange: *validExch,
+		bindings: []string{},
+	}
+	c, err := f.Create()
+	require.NoError(t, err)
+	ctx := context.Background()
+
+	_, _, err = c.Consume(ctx)
+	assert.NoError(t, err)
+	err = c.Close()
+	assert.NoError(t, err)
+}
+
 // Small default publisher for testing purposes.
 func setupRabbitMQPublisher(t *testing.T) *amqp.Channel {
